+++ docs/specs/parts-list-view/design.md
@@ -17,7 +17,7 @@
   - **パーツ一覧ページ**: パーツの探索（フィルタリング、並び替え、お気に入り）に専念
   - **アセンブリページ**: 機体の組み立て（ロック、ランダム、計算結果表示）に専念
 - アセンブリページへのパーツ選択フローが「一覧ページで絞り込み → アセンに遷移 → 母集団から選択」に変更される
-- パーツ一覧ページ専用に再構成したフィルタリングロジック（`packages/web/src/lib/view/parts-list/state/filter`配下）はアセンブリページから切り離され、アセンブリページではフィルタ機能を提供しない
+- パーツ一覧ページ専用に再構成したフィルタリングロジック（`packages/web/src/lib/view/parts-list/filter`配下）はアセンブリページから切り離され、アセンブリページではフィルタ機能を提供しない
 - アセンブリページのナビゲーションバーからパーツ一覧ページ（`/parts-list`）へ遷移する導線を追加し、URLクエリ（フィルタ条件・母集団制限）を保持したままSvelteKitの`goto`で移動できるようにした（失敗時は構造化ログで検知）
 
 ### Goals
