+++ packages/web/src/lib/view/parts-list/PartsListView.spec.ts
@@ -19,12 +19,12 @@ import {
 } from '@testing-library/svelte'
 import { describe, it, expect, beforeEach, afterEach, vi } from 'vitest'
 
-import PartsListView from './PartsListView.svelte'
-import PartsListViewTestWrapper from './PartsListView.test-wrapper.svelte'
 import {
   compressToUrlSafeString,
   decompressFromUrlSafeString,
-} from './state/filter/compression'
+} from './filter/compression'
+import PartsListView from './PartsListView.svelte'
+import PartsListViewTestWrapper from './PartsListView.test-wrapper.svelte'
 import * as partsPoolSerializer from './state/parts-pool-serializer'
 import { sortPartsByKey } from './state/sort'
 import * as stateSerializer from './state/state-serializer'
