+++ packages/web/src/lib/view/parts-list/filter/compression.ts
